pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = [
]
requires-python = ">=3.12,<4"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "discord-py (>=2.6.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "qbittorrent-api (>=2025.7.0,<2026.0.0)",
    "playwright (>=1.57.0,<2.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "flask-cors (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
